Pequenas mudanças

+++ src/controller/userController.ts
@@ -1,5 +1,5 @@
-import { Request, Response } from 'express';
-import { getCustomRepository, getRepository } from 'typeorm';
+import { Request, response, Response } from 'express';
+import { createQueryBuilder, getCustomRepository, getRepository } from 'typeorm';
 import { User } from '../models/User';
 import { UsersRepository } from '../repositories/UsersRepository';
 
@@ -33,11 +33,20 @@ class UserController {
         return res.json(user);
     }
 
-    async delete(req: Request, res: Response) {
-        const { id } = req.body;
+    async delete(req: Request, res: Response, params) {
 
-
+        const { id } = req.params;
         const usersRepository = getCustomRepository(UsersRepository);
+        const userExist = await usersRepository.findOne({
+            id
+        });
+
+        if (!userExist) {
+            return res.json({
+                "message": "User not found"
+            });
+        }
+
         usersRepository.delete({ id });
 
 
@@ -60,6 +69,48 @@ class UserController {
         } else {
             return res.json(user);
         }
+    }
+
+    async update(req: Request, res: Response) {
+        //recebe todos os dados do  usuario a ser editado
+        const { id, name, cpf, email, password, birthDate } = req.body;
+
+        const usersRepository = getCustomRepository(UsersRepository);
+
+        //porem usa apenas o id para localiza-lo no bd
+        let user = await usersRepository.findOne({
+            id
+        });
+
+        if (!user) {
+            return res.status(400).json("User not found");
+        }
+
+
+
+        user.name == name ? user.name = user.name : user.name = name;
+        user.cpf == cpf ? user.cpf = user.cpf : user.cpf = cpf;
+        user.email == email ? user.email = user.email : user.email = email;
+        user.password == password ? user.password = user.password : user.password = password;
+        user.birthDate == birthDate ? user.birthDate = user.birthDate : user.birthDate = birthDate;
+
+
+        await usersRepository.update(id, {
+            name: user.name, cpf: user.cpf, email: user.email,
+            password: user.password, birthDate: user.birthDate
+        })
+
+
+        return res.json(user);
+    }
+
+
+    async show(request: Request, response: Response) {
+        const usersRepository = getCustomRepository(UsersRepository);
+
+        const all = await usersRepository.find();
+
+        return response.json(all);
     }
 
 }
